+++ src/components/EnhancedErrorBoundary.tsx
@@ -1,5 +1,5 @@
 import React, { Component, ErrorInfo, ReactNode } from 'react';
-import { Button } from './ui/button';
+import { Button } from '@/components/ui/Button';
 
 interface Props {
   children: ReactNode;
