+++ src/components/FeaturesGuideSection.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { GradientHeading } from "./GradientHeading";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { Button } from "./ui/button";
+import { Button } from "@/components/ui/Button";
 import { fireEvent } from '@/lib/analytics';
 import { Users, Zap, Settings, MessageSquare, Sparkles, ArrowRight, BarChart3, Plus, HelpCircle } from 'lucide-react';
