+++ src/components/QuoteFormSection.tsx
@@ -1,5 +1,5 @@
 
-import { Button } from "./ui/button";
+import { Button } from "@/components/ui/Button";
 import { Input } from "./ui/input";
 import { GradientHeading } from "./GradientHeading";
 import { useState } from "react";
