+++ src/components/checkout/CheckoutButton.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/ui/Button';
 import { Loader2 } from 'lucide-react';
 
 import { useAuth } from '@/hooks/useAuth';
